feat:整合es+kafka完成房源信息索引

+++ src/main/java/com/lsd/eshouse/common/dto/HouseDTO.java
@@ -28,7 +28,7 @@ public class HouseDTO implements Serializable {
 
     private int floor;
 
-    private Long adminId;
+    private Integer adminId;
 
     private String district;
